+++ RedPandaIDE/widgets/lightfusionstyle.cpp
@@ -11,6 +11,9 @@ int LightFusionStyle::pixelMetric(PixelMetric metric, const QStyleOption *option
     switch ( metric ) {
     case QStyle::PM_SmallIconSize:
         return pointToPixel(pSettings->environment().interfaceFontSize());
+    case QStyle::PM_TabCloseIndicatorHeight:
+    case QStyle::PM_TabCloseIndicatorWidth:
+        return 1.2*pointToPixel(pSettings->environment().interfaceFontSize());
     default:
         return QProxyStyle::pixelMetric( metric, option, widget );
     }